Module and Pkg use TypeAlias
